Add function to print to uart + fix CAN clock

+++ Projet_STM32/CarteCommandeTricycle/Core/Inc/ultrasound.h
@@ -1,6 +1,5 @@
-#include <string.h>
-#include <stdio.h>
+#include <stdint.h>
 
 void usDelay(uint32_t uSec);
-float getDistance();
-float getDistance_TRIG_INPUT();
+double getDistance();
+double getDistance_TRIG_INPUT();
